==> Programas/1 - TFM - Citi Bike System Analysis - READ.Rmd <==
---
title: "TFM - Citi Bike System Analysis"
author: "franmarq@gmail.com"
date: "2024-11-09"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## R Markdown


```{r prepdata}
library(openxlsx)
library(tidyverse)
library(lubridate)

tiempo_inicio <- proc.time()

# Definir la ruta de la carpeta
ruta_carpeta <- "C:/Users/franm/OneDrive/Documents/Personales/Javier/Academicos/UGR - Estadistica Aplicada/Materias/24-25/TFM/data/citybike/"


# Generar una lista de todos los archivos CSV en la carpeta
archivos_csv <- list.files(path = ruta_carpeta, pattern = "\\.csv$")



# Función para leer y filtrar cada archivo
leer_y_filtrar <- function(archivo) {
  ruta_completa <- file.path(ruta_carpeta, archivo)
  read_csv(ruta_completa, col_names = TRUE) %>%
    select(rideable_type, started_at, ended_at, start_station_name, end_station_name,start_lat,start_lng,end_lat,end_lng, member_casual) %>%
    filter(start_station_name %in% c("Central Park S & 6 Ave",
                                    "E 41 St & Madison Ave (SE corner)",
                                    "Liberty St & Broadway",
                                    "W 116 St & Amsterdam Ave",
                                    "Mercer St & Spring St",
                                    "W 40 St & 7 Ave"))
}

# Función para leer y filtrar cada archivo
leer_y_filtrar2 <- function(archivo) {
  ruta_completa <- file.path(ruta_carpeta, archivo)
  read_csv(ruta_completa, col_names = TRUE) %>%
    select(rideable_type, started_at, ended_at, start_station_name, end_station_name,start_lat,start_lng,end_lat,end_lng, member_casual) %>%
    filter(end_station_name %in% c("Central Park S & 6 Ave",
                                    "E 41 St & Madison Ave (SE corner)",
                                    "Liberty St & Broadway",
                                    "W 116 St & Amsterdam Ave",
                                    "Mercer St & Spring St",
                                   "W 40 St & 7 Ave"))
}


# Aplicar la función a cada archivo y combinar los resultados
datos_filtrados <- map_df(archivos_csv, leer_y_filtrar)
datos_filtrados2 <- map_df(archivos_csv, leer_y_filtrar2)


# Mostrar los primeros registros para verificar
#head(datos_filtrados)


```


```{r processing}
# Convertir started_at y ended_at a objetos de fecha y hora
datos_filtrados <- datos_filtrados %>%
  mutate(
    started_at = ymd_hms(started_at),
    ended_at = ymd_hms(ended_at),
    fecha_ride =   as_date(started_at),
    duracion_min = as.numeric(round(difftime(ended_at, started_at, units = "mins"), 2))
  )

datos_filtrados2 <- datos_filtrados2 %>%
  mutate(
    started_at = ymd_hms(started_at),
    ended_at = ymd_hms(ended_at),
    fecha_ride =   as_date(started_at),
    duracion_min = as.numeric(round(difftime(ended_at, started_at, units = "mins"), 2))
  )



```



```{r writerds}

ruta_salida <- "C:/Users/franm/OneDrive/Documents/Personales/Javier/Academicos/UGR - Estadistica Aplicada/Materias/24-25/TFM/data/citybike/datos_filtrados.rds"

ruta_salida2 <- "C:/Users/franm/OneDrive/Documents/Personales/Javier/Academicos/UGR - Estadistica Aplicada/Materias/24-25/TFM/data/citybike/datos_filtrados2.rds"



# Guardar los datos filtrados en formato RData
saveRDS(datos_filtrados, file = ruta_salida)
saveRDS(datos_filtrados2, file = ruta_salida2)

# Mensaje de confirmaciÃ³n
#print("Los datos filtrados se han guardado en:", ruta_salida2)

tiempo_fin <- proc.time()

tiempo_transcurrido <- tiempo_fin - tiempo_inicio

# 4. Imprime el resultado
print(tiempo_inicio)
print(tiempo_fin)
print(tiempo_transcurrido)
```


```{r readWeatherBQ}

#datos del clima 


# Cargar el paquete cada vez que inicies R
library(bigrquery)


# Autenticar tu sesión
bq_auth()


# 1. ---- Define el ID de tu proyecto en GCP ----
project_id <- "api-public-data-access"

# 2. ---- Define la consulta SQL ----
sql_query <- "
SELECT *
FROM `bigquery-public-data.noaa_gsod.gsod20*`
WHERE stn = '725053'
  -- 1. Limita las tablas a escanear solo a los años relevantes
  AND _TABLE_SUFFIX BETWEEN '22' AND '25'
  -- 2. Crea una fecha y filtra por el rango exacto
  AND PARSE_DATE('%Y%m%d', CONCAT(year, mo, da)) BETWEEN '2022-10-01' AND '2025-09-30'
ORDER BY year, mo, da
limit 10000
"

# 3. ---- Ejecuta la consulta ----
bq_job <- bq_project_query(project_id, sql_query)


# 4. ---- Descarga los resultados a un data frame de R ----
datos_clima_nyc <- bq_table_download(bq_job)


datos_clima_nyc%>%
  glimpse()

# Guardar los datos filtrados en formato RDS


file_weather<- "C:/Users/franm/OneDrive/Documents/Personales/Javier/Academicos/UGR - Estadistica Aplicada/Materias/24-25/TFM/data/datos_clima_nyc.rds"



saveRDS(datos_clima_nyc, file = file_weather)

```
