==> Programas/2 - TFM - Citi Bike System Analysis - Read the data source.Rmd <==
---
title: "TFM - Citi Bike System Analysis"
author: "franmarq@gmail.com"
date: "2024-11-09"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#library(openxlsx)
library(tidyverse)
library(lubridate)
library(summarytools)
library(knitr)
library(geosphere)

```

configuramos el directorio local de trabajo. 

Cargamos las librerias de trabajo e Importamos la data previamente ubicada en el folder /data


```{r ReadDatacitibike}

# LECTURA DATOS CITYBIKE:

setwd('C:/Users/franm/OneDrive/Documents/Personales/Javier/Academicos/UGR - Estadistica Aplicada/Materias/24-25/TFM')

# Leer el archivo RDS
data_citybike <- "data/citybike/datos_filtrados.rds" #datos de salidas desde la estacion
data_citybike2 <- "data/citybike/datos_filtrados2.rds" #datos de llegadas a la estacion

datos_rds <- readRDS(data_citybike)
datos_rds2 <- readRDS(data_citybike2)

```

Aplicamos algunes ajustes iniciales al conjunto de forma de no tener problemas asociados a datos faltantes, incongruentes, fuera de rango   

```{r ProcDataCitibike}

# Convertir started_at y ended_at a objetos de fecha y hora

datos_rds$dir<-'salida'
datos_rds2$dir<-'entrada'

datos_rds$date<-datos_rds$fecha_ride
datos_rds2$date<-datos_rds2$fecha_ride

datos_rds$fecha_ride<-NULL
datos_rds2$fecha_ride<-NULL


#fusion datos entrada salida
datos_citibike <- bind_rows(datos_rds, datos_rds2)

rm(datos_rds)
rm(datos_rds2)

#No datos faltantes
datos_citibike<-na.omit(datos_citibike)

#adjustments by imputation values
datos_citibike$duracion_min[datos_citibike$duracion_min < 0]<-0
#datos_citibike$duracion_min[datos_citibike$duracion_min > 60]<-60



datos_citibike$date[datos_citibike$date<'2022-11-01']<-'2022-11-01'
datos_citibike$started_at[datos_citibike$started_at<'2022-11-01']<-'2022-11-01'


#calculo de algunas variables
datos_citibike <- datos_citibike %>%
  mutate(
    started_at = ymd_hms(started_at),
    ended_at = ymd_hms(ended_at),
    hora_ini = hour(started_at),
    hora_fin = hour(ended_at),
    date =   as_date(started_at),
    duracion = as.numeric(round(difftime(ended_at, started_at, units = "mins"), 2))
  )

#Misma ubicacion
datos_citibike$same_location<- ifelse(datos_citibike$start_lat == datos_citibike$end_lat & datos_citibike$start_lng==datos_citibike$end_lng, 1, 0)

#Distancia mts origen/destino 
datos_citibike <- datos_citibike %>%
  mutate(
    distancia = as.integer(distHaversine(
      # Punto 1: (Lon, Lat)
      cbind(start_lng, start_lat),
      # Punto 2: (Lon, Lat)
      cbind(end_lng, end_lat)
    )
  ))


```


```{r ReadDataClima}

# DATOS CLIMA: Seleccion de las variables 
setwd('C:/Users/franm/OneDrive/Documents/Personales/Javier/Academicos/UGR - Estadistica Aplicada/Materias/24-25/TFM')

Outfile_clima_nyc_rds<- "data/datos_clima_nyc.rds"

datos_clima <- readRDS(Outfile_clima_nyc_rds)

```


```{r ProcDataClima}
#Niveles de severidad por la cantidad de Lluvia

niveles_lluvia <- c("Sin Lluvia", 
                       "Ligera", 
                       "Moderada", 
                       "Fuerte", 
                       "Dato Faltante")
niveles_viento <- c("Sin viento", 
                       "Ligero", 
                       "Fuerte", 
                       "Dato Faltante")
niveles_temp <- c("Frio","Templada","Caliente","Dato Faltante")



datos_clima <- datos_clima %>%
#  select(date,temp, visib,wdsp,max,min,prcp,fog,rain_drizzle,snow_ice_pellets,hail,thunder,
#         tornado_funnel_cloud) %>%
  mutate(across(c(fog, rain_drizzle, snow_ice_pellets, hail, 
                  thunder, tornado_funnel_cloud, wdsp, mxpsd), 
                as.numeric)) %>%
  mutate(mal_clima = fog+rain_drizzle+snow_ice_pellets+hail+thunder+tornado_funnel_cloud)%>%
  mutate(
      lluvia_nivel = case_when(
        # GSOD usa 99.99 para datos faltantes
        prcp >= 99             ~ "Dato Faltante", 
        # Lógica de clasificación de 24 horas
        prcp == 0              ~ "Sin Lluvia",
        prcp < 0.40            ~ "Ligera",
        prcp < 2.00            ~ "Moderada",
        TRUE                   ~ "Fuerte" 
        ),# 'TRUE' actúa como el 'else' final (cualquier cosa >= 4.00)
      lluvia_nivel = factor(lluvia_nivel, levels = niveles_lluvia)
    )%>%
  mutate(
      viento_nivel = case_when(
        # GSOD usa 99.99 para datos faltantes
        mxpsd >= 99             ~ "Dato Faltante", 
        # Lógica de clasificación de 24 horas
        mxpsd == 0              ~ "Sin viento",
        mxpsd < 20               ~ "Ligero",
        TRUE                   ~ "Fuerte" 
        ),# 'TRUE' actúa como el 'else' final (cualquier cosa >= 4.00)
      viento_nivel = factor(viento_nivel, levels = niveles_viento)
    )%>%
  mutate(
      temp_nivel = case_when(
        max >= 9999             ~ "Dato Faltante", 
        max < 50.0              ~ "Frio",
        max < 80.0               ~ "Templada",
        TRUE                   ~ "Calor" 
        ),
      temp_nivel = factor(temp_nivel, levels = niveles_temp)
    )

  
datos_clima$mal_clima<-ifelse(datos_clima$prcp>0 & datos_clima$mal_clima==0,1,datos_clima$mal_clima)

plot(datos_clima$date,datos_clima$prcp)

#velocidad del viento
#datos_clima$wdsp<-as.numeric(datos_clima$wdsp)
#datos_clima$mxpsd<-as.numeric(datos_clima$mxpsd)

plot(datos_clima$date,datos_clima$wdsp)

hist(datos_clima$wdsp)
hist(datos_clima$mxpsd)
hist(datos_clima$max)



freq(datos_clima$lluvia_nivel)
freq(datos_clima$viento_nivel)
freq(datos_clima$temp_nivel)
freq(datos_clima$max)

mi_tabla_base <- table(datos_clima$lluvia_nivel, datos_clima$mal_clima)

# Imprimir la tabla
print(mi_tabla_base)

datos_clima2 <- datos_clima %>%
  select(date,max,prcp,mxpsd,mal_clima,lluvia_nivel, viento_nivel, temp_nivel)

```

```{r fusion}

datos_citibike2 <- left_join(datos_citibike, datos_clima2, by = "date")

```





```{r ReportCitibike}
## REPORTES CITIBIKE


datos_rds%>%
  glimpse()

datos_rds2%>%
  glimpse()


# after Results
datos_rds_sum_tbl <- datos_rds %>%
dfSummary(plain.ascii = FALSE, style = "grid", graph.magnif = 0.75, method="render")


datos_rds_sum_tbl%>%
    select(-text.graph) %>%
    knitr::kable(align = "c")

```

Inicialmente tenemos 8 variables, de las cuales 






```{r ReportDataclima}

datos_clima_nyc2%>%
  glimpse()

# Crea la tabla de frecuencia
tabla_frecuencias <- datos_clima_nyc2 %>%
  count(bad_weather, sort = TRUE, name = "frecuencia") %>%
  mutate(
    porcentaje_num = round((frecuencia / sum(frecuencia)) * 100, 2)
    )

datos_clima_nyc2 <- datos_clima_nyc2 %>%
  select(-fog,-rain_drizzle,-snow_ice_pellets,-hail,-thunder,-tornado_funnel_cloud)


# Muestra la tabla en la consola
print(tabla_frecuencias)



```
